Safari-611.1.10/LayoutTests/fast/multicol/vertical-lr/standardized-break-properties-expected.txt

PASS: 'break-before' is at (8, 220)
PASS: 'after-break' is at (8, 431)
